requirements.txt
appdirs==1.4.3
arrow==0.13.0
astroid==2.1.0
attrs==18.2.0
backcall==0.1.0
black==19.3b
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.3.2
dj-database-url==0.5.0
Django==2.1.4
django-extensions==2.1.4
django-filter==2.1.0
django-heroku==0.3.1
django-open-humans==0.1.2.3
django-simple-bulma==1.1.4
djangorestframework==3.9.1
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
humanfriendly==4.17
idna==2.8
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
libsass==0.17.0
Markdown==3.0.1
mccabe==0.6.1
open-humans-api==0.2.5
parso==0.3.3
pickleshare==0.7.5
prompt-toolkit==2.0.8
psycopg2==2.7.6.1
Pygments==2.3.1
pyreadline==2.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
six==1.12.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.0
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1